Delete a model and its data

Delete a model from your local machine that you downloaded using the pull() function. To see which models are available, use the list_models() function.

Usage

delete(name, endpoint = "/api/delete", host = NULL)

Arguments

name

A character string of the model name such as "llama3".

endpoint

The endpoint to delete the model. Default is "/api/delete".

host

The base URL to use. Default is NULL, which uses Ollama's default base URL.

Value

A httr2 response object.

References

API documentation

Examples

if (FALSE) { # \dontrun{
delete("llama3")
} # }