Search for options based on a query

Search for options based on a query

Usage

search_options(query)

Arguments

query

A query (character) to search for in the options.

Value

Returns a list of matching options.

Examples

search_options("learning rate")
#> Matching options: mirostat_eta 
#> $mirostat_eta
#> $mirostat_eta$description
#> [1] "Influences how quickly the algorithm responds to feedback from the generated text. A lower learning rate will result in slower adjustments, while a higher learning rate will make the algorithm more responsive."
#> 
#> $mirostat_eta$default_value
#> [1] 0.1
#> 
#> 
search_options("tokens")
#> Matching options: tfs_z, num_predict 
#> $tfs_z
#> $tfs_z$description
#> [1] "Tail free sampling is used to reduce the impact of less probable tokens from the output. A higher value (e.g., 2.0) will reduce the impact more, while a value of 1.0 disables this setting."
#> 
#> $tfs_z$default_value
#> [1] 1
#> 
#> 
#> $num_predict
#> $num_predict$description
#> [1] "Maximum number of tokens to predict when generating text. (Default: 128, -1 = infinite generation, -2 = fill context)"
#> 
#> $num_predict$default_value
#> [1] 128
#> 
#> 
search_options("invalid query")
#> No matching options found
#> 
#> list()